removed globally defined default icon set
